OSQuery file support plus new data
