+++ src/rust/generators/osquery-subgraph-generator/src/parsers/grapl_pack/mod.rs
@@ -1,0 +1,16 @@
+use serde::{Serialize, Deserialize, Deserializer};
+use std::fmt::Display;
+use std::str::FromStr;
+
+pub mod processes;
+mod files;
+
+fn from_str<'de, D, T>(deserializer: D) -> Result<T, D::Error>
+    where
+        T: FromStr,
+        T::Err: Display,
+        D: Deserializer<'de>,
+{
+    let s = String::deserialize(deserializer)?;
+    T::from_str(&s).map_err(serde::de::Error::custom)
+}
