--- src/rust/generators/osquery-subgraph-generator/src/parsers/packs/mod.rs
@@ -1,2 +1,0 @@
-use serde::{Serialize, Deserialize};
-pub mod grapl;
